+++ api/routes/coinsRoutes.js
@@ -1,12 +1,12 @@
 const { Router } = require('hyper-express');
-const coinsController = require('../controllers/coinsController');
+const coins = require('../controllers/coins');
 
 const router = new Router();
 
-router.get('/metadata', coinsController.getCoinsMetadata);
-router.get('/current', coinsController.getCoinsCurrent);
-router.get('/timeseries', coinsController.getCoinsTimeseries);
-router.get('/earliest', coinsController.getCoinFirstTimestamp);
-router.get('/percentage-change', coinsController.getPercentageChange);
+router.get('/metadata', coins.getCoinsMetadata);
+router.get('/current', coins.getCoinsCurrent);
+router.get('/timeseries', coins.getCoinsTimeseries);
+router.get('/earliest', coins.getCoinsFirstTimestamp);
+router.get('/percentage-change', coins.getCoinsPercentageChange);
 
 module.exports = router;
