+++ utils/index.js
@@ -19,7 +19,40 @@ async function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms))
 }
 
+function parsePidMapping(pidInput) {
+  let inputArray = [];
+  if (typeof pidInput === 'string') {
+    inputArray = pidInput.split(',').map(p => p.trim()).filter(Boolean);
+  } else if (Array.isArray(pidInput)) {
+    inputArray = pidInput.map(p => p.trim()).filter(Boolean);
+  }
+  if (inputArray.length > 100) {
+    throw new Error("Maximum of 100 PID tokens allowed.");
+  }
+  const mapping = {};
+  const normalizedPids = inputArray.map(pid => {
+    const normalized = normalizeCoinId(pid);
+    mapping[pid] = normalized;
+    return normalized;
+  });
+  return { mapping, normalizedPids };
+}
+
+function getEffectivePids(originalPid, mapping, metadata) {
+  const normalized = mapping[originalPid];
+  let candidates = [];
+  const meta = metadata[originalPid];
+  if (meta && Array.isArray(meta.redirects) && meta.redirects.length > 0) {
+    candidates = [...meta.redirects];
+  }
+  candidates.push(normalized);
+  return [...new Set(candidates)];
+}
+
+
 module.exports = {
   normalizeCoinId,
   sleep,
+  parsePidMapping,
+  getEffectivePids
 }
